+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/api/dto/EntityFactory.java
@@ -96,10 +96,6 @@ public final class EntityFactory {
         return entity;
     }
 
-    private void pairDownDiagnostics(final ProcessorDiagnosticsDTO dto, final PermissionsDTO controllerPermissions) {
-
-    }
-
     public StatusHistoryEntity createStatusHistoryEntity(final StatusHistoryDTO statusHistory, final PermissionsDTO permissions) {
         final StatusHistoryEntity entity = new StatusHistoryEntity();
         entity.setCanRead(permissions.getCanRead());
